Fixing The Expiry Error
There was an issue when it came to expiry value being represented as a floating number when it came to the newer chrome driver versions. To resolve this problem I added an if statement that would check if expiry was an instance of a floating number and will convert it to an integer so that the driver.add_cookie(cookie) can accept it since it only accepts int values not floating values.  This is only a change in the load_cookies function since that was the one I seemed to have the only problem with.  Hopefully, this is of some help learned a lot from watching your youtube video on this.